+++ lib/src/rules/ntlife.rs
@@ -301,7 +301,10 @@ impl Rule for NtLife {
         cell.desc.set(desc);
     }
 
-    fn consistify<'a>(world: &mut World<'a, Self>, cell: CellRef<'a, Self>) -> bool {
+    fn consistify<'a, RE: Reason<'a, Self>>(
+        world: &mut World<'a, Self, RE>,
+        cell: CellRef<'a, Self>,
+    ) -> bool {
         let flags = world.rule.impl_table[cell.desc.get().0 as usize];
 
         if flags.is_empty() {
@@ -319,7 +322,7 @@ impl Rule for NtLife {
                 ALIVE
             };
             let succ = cell.succ.unwrap();
-            return world.set_cell(succ, state, Reason::Deduce);
+            return world.set_cell(succ, state, RE::from_cell(cell));
         }
 
         if flags.intersects(ImplFlags::SELF) {
@@ -328,7 +331,7 @@ impl Rule for NtLife {
             } else {
                 ALIVE
             };
-            if !world.set_cell(cell, state, Reason::Deduce) {
+            if !world.set_cell(cell, state, RE::from_cell(cell)) {
                 return false;
             }
         }
@@ -345,7 +348,7 @@ impl Rule for NtLife {
                             } else {
                                 ALIVE
                             };
-                            if !world.set_cell(neigh, state, Reason::Deduce) {
+                            if !world.set_cell(neigh, state, RE::from_cell(cell)) {
                                 return false;
                             }
                         }
@@ -474,7 +477,10 @@ impl Rule for NtLifeGen {
         cell.desc.set(desc);
     }
 
-    fn consistify<'a>(world: &mut World<'a, Self>, cell: CellRef<'a, Self>) -> bool {
+    fn consistify<'a, RE: Reason<'a, Self>>(
+        world: &mut World<'a, Self, RE>,
+        cell: CellRef<'a, Self>,
+    ) -> bool {
         let desc = cell.desc.get();
         let flags = world.rule.impl_table[desc.0 as usize];
         let gen = world.rule.gen;
@@ -494,7 +500,7 @@ impl Rule for NtLifeGen {
                         ALIVE
                     };
                     let succ = cell.succ.unwrap();
-                    return world.set_cell(succ, state, Reason::Deduce);
+                    return world.set_cell(succ, state, RE::from_cell(cell));
                 }
             }
             Some(ALIVE) => {
@@ -510,7 +516,7 @@ impl Rule for NtLifeGen {
                         ALIVE
                     };
                     let succ = cell.succ.unwrap();
-                    return world.set_cell(succ, state, Reason::Deduce);
+                    return world.set_cell(succ, state, RE::from_cell(cell));
                 }
             }
             Some(State(i)) => {
@@ -518,13 +524,13 @@ impl Rule for NtLifeGen {
                     return j == (i + 1) % gen;
                 } else {
                     let succ = cell.succ.unwrap();
-                    return world.set_cell(succ, State((i + 1) % gen), Reason::Deduce);
+                    return world.set_cell(succ, State((i + 1) % gen), RE::from_cell(cell));
                 }
             }
             None => match desc.1 {
                 Some(DEAD) => {
                     if flags.contains(ImplFlags::SELF_ALIVE) {
-                        return world.set_cell(cell, State(gen - 1), Reason::Deduce);
+                        return world.set_cell(cell, State(gen - 1), RE::from_cell(cell));
                     } else {
                         return true;
                     }
@@ -536,13 +542,13 @@ impl Rule for NtLifeGen {
                         } else {
                             ALIVE
                         };
-                        if !world.set_cell(cell, state, Reason::Deduce) {
+                        if !world.set_cell(cell, state, RE::from_cell(cell)) {
                             return false;
                         }
                     }
                 }
                 Some(State(j)) => {
-                    return world.set_cell(cell, State(j - 1), Reason::Deduce);
+                    return world.set_cell(cell, State(j - 1), RE::from_cell(cell));
                 }
                 None => return true,
             },
@@ -560,7 +566,7 @@ impl Rule for NtLifeGen {
             for (i, &neigh) in cell.nbhd.iter().enumerate() {
                 if flags.intersects(ImplFlags::from_bits(1 << (2 * i + 6)).unwrap()) {
                     if let Some(neigh) = neigh {
-                        if !world.set_cell(neigh, ALIVE, Reason::Deduce) {
+                        if !world.set_cell(neigh, ALIVE, RE::from_cell(cell)) {
                             return false;
                         }
                     }